+++ lib/Perl/LanguageServer/DebuggerProcess.pm
@@ -4,6 +4,8 @@ use 5.006;
 use strict;
 use Moose ;
 
+use Encode::Locale;
+use Encode;
 use File::Basename ;
 use Coro ;
 use Coro::AIO ;
@@ -25,7 +27,7 @@ has 'program' =>
 
 has 'args' =>
     (
-    isa => 'ArrayRef',
+    isa => 'ArrayRef | Str',
     is  => 'ro',
     default => sub { [] },
     ) ;
@@ -41,6 +43,18 @@ has 'cwd' =>
     (
     isa => 'Maybe[Str]',
     is  => 'ro',
+    ) ;
+
+has 'sudo_user' =>
+    (
+    isa => 'Maybe[Str]',
+    is  => 'ro',
+    ) ; 
+
+has 'path_map' =>
+    (
+    isa => 'Maybe[ArrayRef]',
+    is  => 'rw'
     ) ;
 
 has 'stop_on_entry' =>
@@ -91,6 +105,25 @@ sub BUILDARGS
     $args -> {reload_modules} = delete $args -> {reloadModules}?1:0 ;
     $args -> {stop_on_entry} = delete $args -> {stopOnEntry}?1:0 ;
     $args -> {session_id}    = delete $args -> {__sessionId} || $session_cnt ;
+    $args -> {sudo_user}    = delete $args -> {sudoUser} ;
+    my $map   = delete $args -> {pathMap} ;
+    if ($map)
+        {
+        my $fn ;
+        foreach (@$map)
+            {
+            $fn = $_ -> [0] ;
+            $fn =~ s/^file:// ;
+            $fn =~ s/^\/\/\//\// ;
+            $_ -> [2] ||= $fn ;
+            $fn = $_ -> [1] ;
+            $fn =~ s/^file:// ;
+            $fn =~ s/^\/\/\//\// ;
+            $_ -> [3] ||= $fn ;
+            }
+        $args -> {path_map} = $map ;
+        }
+
     $session_cnt++ ;
 
     return $args ;
@@ -103,6 +136,43 @@ sub logger
     my $self = shift ;
 
     $self -> debug_adapter -> logger (@_) ;
+    }
+
+# ---------------------------------------------------------------------------
+
+sub file_server2client
+    {
+    my ($self, $workspace, $fn) = @_ ;
+
+    return $workspace -> file_server2client ($fn, $self -> path_map) ;
+    }
+
+# ---------------------------------------------------------------------------
+
+sub file_client2server
+    {
+    my ($self, $workspace, $fn) = @_ ;
+
+    return $workspace -> file_client2server ($fn, $self -> path_map) ;
+    }
+
+# ---------------------------------------------------------------------------
+
+sub add_path_mapping
+    {
+    my ($self, $fn_server, $fn_client) = @_ ;
+    my $map = $self -> path_map ;
+    $map = $self -> path_map ([]) if (!$map) ;
+
+
+    foreach my $m (@$map)
+        {
+        #print STDERR "add file_server2client $m->[2] -> $m->[3]\n" ;
+        return if ($fn_server eq $m->[2]) ;
+        }
+
+    unshift @$map, ['file://' . $fn_server, 'file://' . $fn_client, $fn_server, $fn_client] ;
+    return  ;
     }
 
 # ---------------------------------------------------------------------------
@@ -120,13 +190,19 @@ sub launch
     {
     my ($self, $workspace, $cmd) = @_ ;
 
-    my $fn = $workspace -> file_client2server ($self -> program) ;
+    my $fn = $self -> file_client2server ($workspace, $self -> program) ;
     my $pid ;
     {
     local %ENV = %ENV ;
+    my @sudoargs ;
+    if ($self->sudo_user)
+        {
+        push @sudoargs, "sudo", "-u", $self->sudo_user ;
+        }
     foreach (keys %{$self -> env})
         {
         $ENV{$_} = $self -> env -> {$_} ;
+        push @sudoargs, "$_=" . $self -> env -> {$_} if $self->sudo_user;
         }
 
     my $cwd ;
@@ -145,7 +221,22 @@ sub launch
     $ENV{PLSDI_OPTIONS} = $self -> reload_modules?'reload_modules':'' ;
     $ENV{PERL5DB}      = 'BEGIN { $| = 1 ; ' . $cwd . 'require Perl::LanguageServer::DebuggerInterface; DB::DB(); }' ;
     $ENV{PLSDI_SESSION}= $self -> session_id ;
-    $pid = $self -> run_async ([$cmd, @inc, '-d', $fn, @{$self -> args}]) ;
+    if ($self->sudo_user)
+        {
+        push @sudoargs, "PLSDI_REMOTE=$ENV{PLSDI_REMOTE}" ;
+        push @sudoargs, "PLSDI_OPTIONS=$ENV{PLSDI_OPTIONS}" ;
+        push @sudoargs, "PERL5DB=$ENV{PERL5DB}" ;
+        push @sudoargs, "PLSDI_SESSION=$ENV{PLSDI_SESSION}" ;
+        }
+
+    if (ref $self -> args)       # ref is array
+        {
+        $pid = $self -> run_async ([@sudoargs, $cmd, @inc, '-d', $fn, @{$self -> args}]) ;
+        } 
+    else                      # no ref is string
+        {
+        $pid = $self -> run_async (join (' ', @sudoargs, $cmd, @inc, '-d', $fn, $self -> args)) ;
+        }
     }
 
     $self -> pid ($pid) ;
@@ -181,6 +272,7 @@ sub on_stdout
 
     foreach my $line (split /\r?\n/, $data)
         {
+        $line = decode(locale => $line);
         $self -> send_event ('output', { category => 'stdout', output => $line . "\r\n" }) ;
         }
     }
@@ -193,6 +285,7 @@ sub on_stderr
 
     foreach my $line (split /\r?\n/, $data)
         {
+        $line = decode(locale => $line);
         $self -> send_event ('output', { category => 'stderr', output => $line . "\r\n" }) ;
         }
     }